+++ src/main/java/com/realdolmen/shopr/service/UserService.java
@@ -12,7 +12,7 @@ public class UserService {
     @Inject
     private UserRepository userRepository;
 
-    public User findUserById(int id) {
+    public User findUserById(long id) {
         return userRepository.findById(id);
     }
 
@@ -24,7 +24,7 @@ public class UserService {
         return userRepository.findByName(name);
     }
 
-    public void updateName(int id, String newName) {
+    public void updateName(long id, String newName) {
         User user = userRepository.findById(id);
         user.setName(newName);
     }
@@ -32,7 +32,7 @@ public class UserService {
     public void insert(User user) {
         userRepository.insert(user);
     }
-    public void removeById(int id) {
+    public void removeById(long id) {
         userRepository.remove(id);
     }
